+++ app/src/main/java/com/example/se306_project1/models/IProduct.java
@@ -3,18 +3,18 @@ package com.example.se306_project1.models;
 import java.util.ArrayList;
 
 public interface IProduct {
-    public int getProductID();
+    public long getProductID();
     public Double getProductPrice();
     public String getProductLongName();
     public String getProductShortName();
     public String getProductDescription();
     public String getProductDetails();
     public String getProductCare();
-    public int getCategoryID();
+    public long getCategoryID();
     public String getBrandName();
     public String getProductColourType();
     public Boolean getIsFavourite();
     public void setProductIsFavourite(Boolean isFavourite);
-    public int getProductCountVisit();
+    public long getProductCountVisit();
     public ArrayList<String> getProductImages();
 }
